% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBMWP_CR.r
\name{calcBMWP_CR}
\alias{calcBMWP_CR}
\title{Calculate the Costa Rican BMWP biotic index for freshwater invertebrate samples}
\usage{
calcBMWP_CR(df, type = "num")
}
\arguments{
\item{df}{A dataframe containing list of taxon names and their abundances
in samples, along with sample identifiers.  Default format is for taxon
names to be in the first column and sample abundances in subsequent
columns with identifers as column headers. See built-in \code{\link{almond}}
 dataset for an example. If data are in the transposed format i.e taxa as
 columns and samples as rows, the \code{\link{transposedata}} function can
 be used prior to calculation.}

\item{type}{Indicates type of data being processed. Options are "num" for
numeric data, "log" for integer log abundance categories (1-5) or "alpha"
for alphabetic abundance categories (A-E). Default value is "num".}
}
\value{
A data frame consisting of columns of index values with samples in
rows.
}
\description{
Calculates the Costa Rican version of the BMWP freshwater invertebrate biotic index.
}
\examples{

# calculate the Costa Rican version of the BMWP index for
# the built in 'almond' dataset
# 'type' does not have to specified as default is used
# ("num")

calcBMWP_CR(almond)

# example of processing data in alphabetic log abundance categories
# using the 'type' argument

# 'braidburn' dataset contains alphabetic log category data
# see ?braidburn for details

# calculate the Costa Rican BMWP index for this dataset

calcBMWP_CR(braidburn, "alpha")

# example of processing data in numeric log abundance categories
# using the 'type' argument

# 'greenburn' dataset contains numeric log category data
# see ?greenburn for details

# calculate the Costa Rican BMWP index for this dataset

calcBMWP_CR(greenburn, "log")

}
